src/agent_runtime/requirements.txt
crewai
langchain_openai
psycopg2-binary
pandas
python-dotenv
crewai-tools
langchain-ollama
langchain
langchain-community
litellm
duckduckgo-search
ddgs
scikit-learn
numpy
pillow
reportlab
pytest
requests
watchdog
gitpython
